clearing compiler warnings for 64-bit conversion
git-svn-id: svn://orca.physics.unc.edu/Orca64@11893 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/App Delegate/ORAppDelegate.m
@@ -541,7 +541,7 @@ NSString* OROrcaFinalQuitNotice      = @"OROrcaFinalQuitNotice";
 	
 	[[self undoManager] removeAllActions];
 
-	int     count  = [[ORGlobal sharedGlobal] cpuCount];
+	NSUInteger     count  = [[ORGlobal sharedGlobal] cpuCount];
 	if(count==1){
 		[self closeSplashWindow];
 		NSLogColor([NSColor redColor],@"Number of processors: %d\n",count);
